Added player movement and scrolling
Created a new class 'Player' that controls drawing and moving the player. The background is drawn relative to the player's position, giving a scrolling background. The player also has a hopping animation when they walk. I also made it possible to flip sprites horizontally so the wizard changes which direction they face as they move.