Build support
Made the executable into a wrapper for the actual python code for the project and included all the files needed to run it. Made a simple build script in case I ever need to re-make the exe (though I shouldn't have to). Made highscore code more portable (program doesn't crash if the file doesn't exist yet)